Rename to validator staking desktop
