+++ src/components/validator_staking_desktop/styles.tsx
