+++ src/index.ts
@@ -19,7 +19,7 @@ import ValidatorInfo from './components/validator_info';
 import ActivityDetails from './components/activity_details';
 import InfoPopover from './components/info_popover';
 import PowerEvents from './components/power_events';
-import ValidatorDelegationsDesktop from './components/validator_delegations_desktop';
+import ValidatorStakingDesktop from './components/validator_staking_desktop';
 
 export {
   DataBlock,
@@ -43,6 +43,6 @@ export {
   ActivityDetails,
   InfoPopover,
   PowerEvents,
-  ValidatorDelegationsDesktop,
+  ValidatorStakingDesktop,
 
 };
